fix: remove response-get-value as well, since it was only used once
