+++ src/messages/responses/cache-dictionary-get-field.ts
@@ -1,13 +1,42 @@
 // older versions of node don't have the global util variables https://github.com/nodejs/node/issues/20365
-import * as GetValue from './response-get-value';
 import {TextDecoder} from 'util';
 import {SdkError} from '../../errors/errors';
+import {
+  ResponseBase,
+  ResponseError,
+  ResponseHit,
+  ResponseMiss,
+} from './response-base';
+import {truncateString} from '../../internal/utils/display';
 
 const TEXT_DECODER = new TextDecoder();
 
-export abstract class Response extends GetValue.Response {}
+export abstract class Response extends ResponseBase {}
 
-export class Hit extends GetValue.Hit {
+class _Hit extends Response {
+  private readonly body: Uint8Array;
+  constructor(body: Uint8Array) {
+    super();
+    this.body = body;
+  }
+  /**
+   * decodes the body into a utf-8 string
+   * @returns string
+   */
+  public valueString(): string {
+    return TEXT_DECODER.decode(this.body);
+  }
+
+  public valueUint8Array(): Uint8Array {
+    return this.body;
+  }
+
+  public override toString(): string {
+    const display = truncateString(this.valueString());
+    return `${super.toString()}: ${display}`;
+  }
+}
+export class Hit extends ResponseHit(_Hit) {
   private readonly field: Uint8Array;
 
   constructor(body: Uint8Array, field: Uint8Array) {
@@ -24,7 +53,8 @@ export class Hit extends GetValue.Hit {
   }
 }
 
-export class Miss extends GetValue.Miss {
+class _Miss extends Response {}
+export class Miss extends ResponseMiss(_Miss) {
   private readonly field: Uint8Array;
 
   constructor(field: Uint8Array) {
@@ -41,7 +71,13 @@ export class Miss extends GetValue.Miss {
   }
 }
 
-export class Error extends GetValue.Error {
+class _Error extends Response {
+  constructor(protected _innerException: SdkError) {
+    super();
+  }
+}
+
+export class Error extends ResponseError(_Error) {
   private readonly field: Uint8Array;
 
   constructor(public _innerException: SdkError, field: Uint8Array) {
